refactor(groq-codegen): consolidate utils via factory
